feat: Move CLI functions to respective modules
